gfx: ANSI screengrabs .png->.c textures via GIMP
Cranksy kindly provided a bunch of graphics for this collaboration,
mostly ANSI-derived but we'll treat them as pixel art.

My understanding is this is a mix of ripped stuff and some original
work, it's nothing crazy - just something usable on short notice
for the 24hr IRC compo this is being made for.

I'm committing all the graphics separately so the big code commit isn't
badly polluted by the gfx bulk