game: fix typo in any_aabb initialization
Missing comma silently resulted in aabb_any.min.x == -2

Excellent example of how using the same syntax for negative and
the subtraction operator creates opportunity for silent bugs.

This is the first time this particular thing has bitten me, doing
mostly systems programming throughout my life I haven't dealt much
with negative values and certainly not frequently initializing bags
of them as is common in game development.

This bug is why the collision hit boxes seemed so huge, it was only
on the horizontal axis, and actually just the minimum side but in-play
it appears to be both since there's always a minimum involved on a
horizontal collision.

Sigh.