*: introduce a projection matrix and winmodes
The bulk of this is mechanical wiring up of a projection_x to all
the nodes.

But this also introduces maintenance of the projection_x, with
aspect-ratio preservation in two of the modes: WINDOWED and
FULLSCREEN, with the previously default stretched-to-fill
fullscreen mode now relegated to a third FILLSCREEN winmode.

The default at startup is now an aspect ratio-preserving windowed
mode.

Simply press the 'f' key at any time to cycle through them.

This was mostly done in sars to provide a source-available test
for reproducing a fullscreen SDL2 bug I filed @
https://github.com/libsdl-org/SDL/issues/5139

Otherwise it's pretty silly to bother with doing anything on
sars...  but it is a handy little mule for such things.