sars: default emscripten builds to fullscreen
Since emscripten builds are often run under a browser window,
putting ourselves into fullscreen mode is most likely to get a
filled-out canvas within the window we find ourselves in.

It's really up to the emscripten runtime environment to deal with
windowing vs. fullscreen handling in an appropriate manner.  We
can't really be bothered with that from the nested environment we
find ourselves in there.