game: infect TV-attracted babies vs. static infections
Existing code wouldn't notice when a TV-attracted baby collided
with an existing inanimate infection (previously infected baby)

This commit adds a nested search in tv_search for the affected
baby case, to detect such collisions.  Note this required bumping
the max nested search for the ix2 from 1 to 2, comment updated to
reflect that.