game: introduce toilet paper hoarding mechanic
for now toilet paper is simply picked up by the adult and
accumulates as a pointless icon starting at the top left corner,
potentially filling the entire screen with enough hoarding;
obstructing the player's view of what's going on.

There's no points added for hoarding, it's completely pointless
and counterproductive.