game: add a variety of teepee powerup quantities
In the interests of making it realistically possible to
completely fill the 256 spots of teepee icons as a winning
condition, there's now mega packs of rolls.

There will need to be some visible indication of qty added,
either numerically or with different graphics.  I'm leaning
towards just adding numeric quantity values like +3 or +36
over the roll which floats away when you take it.