sars: chdir() to executable's base path
On MacOS in particular it seems when executing sars via gui from
the downloaded .zip the cwd isn't being set to the containing
directory beforehand.  This prevents the assets from being found
at the correct location.

With this commit SDL_GetBasePath() is used to try find the
executable's containing directory.  When successful, a chdir() is
performed to this path in an effort to ensure the assets/ dir is
where expected in relative terms.

These are all done in a non-fatal best-effort fashion, simply
warning when failed.

An unfortunate side effect of this is when you're explicitly
running sars from somewhere else with a different relative
assets/ directory for development/testing purposes, it no longer
uses the inherited cwd.  Instead it will always move itself to
wherever the executable resides.  Maybe I'll add a flag for
controlling this behavior.

It's also unclear to me how safe assuming chdir() works will be
vs. platforms supported by SDL2.  Sure, chdir() is POSIX, but um,
SDL2 speaks to supporting platforms like PSP and Android.  I
doubt chdir() is available everywhere, and I wonder why there's
no SDL_chdir() in SDL_filesystem.h to hide this potential detail
alongside the SDL_GetBasePath() abstraction.