game: fix flashers list corruption
(entity_t.flashes_remaining == 0) was being misused to indicate
an entity wasn't on any of the flashers lists.

At some point during the feature's development that was actually
true, but as the code evolved that was no longer the case and
entities would sit in the flashers_off list with
flashes_remaining at 0 before flipping back on for a cycle at
which point they'd be found with 0 flashes_remaining and dropped
from any flashers lists.

This is a problem because flash_entity() would come along and
use the flashes_remaining == 0 state to see if the entity needed
to get on a flashers list before resetting the flashes_remaining
count.

Furthermore, when adult->holding gets cleared (baby drop off) the
flashes_remaining gets zeroed to ensure the reset baby doesn't
start out with residual flashing at its new location.  This just
created more opportunities where the flashes_remaining is zero
while still on a list, since the holding reset was just
expediting the list removal - not actually doing it.

This commit simply introduces a discrete flag for flashing vs.
not flashing to indicate flashers_{on,off} list membership,
independent of the flashes_remaining count.

Fixes infinite loop / apparent freeze bug reported by Tara and
Phil during some play testing.  Thanks guys!