sars: only request a real GLES2 context on emscripten
On MacOS the window create can't even succeed when requesting the
GLES2 context.

Seeing as this GLES2 transition was done just to get emscripten
support, let's limit requesting an actual GLES2 context to the
emscripten builds, and still request the same old GL2.1 context
as pre-emscripten for everything else.

It's unclear to me if this will actually work on win/macos, since
the code is still all targeting the GLES2 API - even the glad
abstraction is the GLES2 variant.  But the thing is, GLES2 API is
largely just a subset of GL2.1, in some ways it's a cleanup of
GL2.1 like removing a bunch of the fixed function stuff I never
used anyways.  The main annoyance with GLES2 is the shader
language version is older than GL2.1 supports.  But if I can live
with constraining my API/shader usage to GLES2 land, it might
just work on GL2.1 contexts without modifications.