game: respawn babies on a timer, and detect immediate infections
Previously respawned babies could respawn on a virus without
becoming infected, and this actually had a calming effect on the
late stage of the game which is the opposite of what should be
happening as the field fills up with viruses.

What would happen is eventually most of the active set of babies
would be sitting on viruses and isolated by viruses so they
wouldn't likely be resuced, but they also weren't becoming
infected despite being on viruses.

With this change, baby respawning is handled strictly by a timer,
currently every .5s.  In that timer handler the number of babies
off from the GAME_NUM_BABIES (10 presently) are attempted to be
respawned.  Any of those which are immediately respawned onto
viruses are immediately infected.  Then the same thing is
revisted on the next timer expiration.

So it doesn't try very hard to maintain GAME_NUM_BABIES as alive
at all times - this actually becomes impossible when the field is
too crowded to even find spots to respawn without infection.
Instead, what will happen, is in the end game you just have
constant babies respawning and dying every .5s.

The gameplay is dramatically different now, and it's potentially
too difficult now to actually manage to hoard 256 rolls of
TP before dying.  I haven't managed it in my testing so far...
so the probabilities may need to be revisited again.

However, the end game /feels/ much more _right_ after this
change.  You end up with just tons of viruses and death, and if
you've managed to hoard a bunch of toilet paper, it's a grim
satirical scene of zero living babies, viruses everywhere, but a
happy pile of TP icons.  I can only imagine how ridiculous it
looks if you manage to actually win and get the animated
celebratory wall of TP over the sea of death.

Now the dead babies turned viruses just need their own unique
sprite showing actual infected corpses instead of the same virus
sprite.  So it's actually visibly a sea of infected bodies...