assets: fix mask.ans to include SAUCE metadata
Dunno why PabloDraw omitted the SAUCE info when creating this
originally, but it was producing an 80 column rendering of the
mask which is more like 43 columns.  This resulted in the
collision box/aabb being far larger than the actual mask as
rendered.

Now with the fixed rendering the mask's scale factor needed some
tweaking as well to preserve the same visual size more or less.