game: tweak teepee probabilities a bit again
Now that respawns can get immediately infected, things get crazy
as they should - so make it a bit easier to hoard all the teepee.

I'm pretty sure you can just hide off-screen at the top and
collect teepee and masks without ever touching a baby and just
dodging animated viruses as a way to cheat and "win".  But that's
kind of allegorical for the pandemic, isn't it?  Ignore the needs
of others while hoarding teepee and masks FTW!