game: differentiate corpse viruses
tv-attracted babies that contact non-corpse viruses should reset
the virus, but they never reset viruses until now.

This commit introduces a corpse flag to the virus entity and sets
it on infection-produced viruses.

Now the tv-attracted baby<->virus infection resets the virus when
the virus isn't a corpse, instead of letting the animated
virus continue on unimpeded.

The corpses should be allowed to propagate infections to
eachother while accumulating, hence why it can't be an
indiscriminate reset of the virus.