game: shorten game over (loss) delay
The delay primarily serves to prevent accidental keypresses from
the gameplay at death bleeding into the high score wait state,
which would immediately leave the high score.

But a full second delay on the losing game over is kind of long,
esp. when you're agitated and just want to immediately restart.

The delay is also overloaded a bit in the winning game over
state, as it governs the TP explosion duration which occurs
within this debouncing-like delay.  So this commit diverges the
two and leaves the winning delay at 1s, while turning the losing
delay to .5s.