game: introduce MAGA powerup and MAGA mode for adult
The MAGA powerup appears only when the adult is captivated by the
TV.  It's technically not guaranteed to occur, but the
propability is so high currently it's basically always going to
appear when captivated.

The current maga.ans is a MAGA hat, it comes up slowly but
otherwise identically to the others like mask/tp etc.

If you take the MAGA hat, the adult puts it on, discards any mask
it's wearing, and the controls become inverted.  Any subsequent
mask pickups attempted simply discard the mask and continue
maskless.  Once in MAGA mode, there is no exiting MAGA mode until
death.