plasma-node,shader-node: break out maga plasma shader
Performing the maga branch in the shader slows things down
considerably.

This commit gets rid of the in-shader branch by splitting the
plasma fragment shader into maga and non-maga variants.

This required adding shader vector support to shader-node so
multiple shaders could be supplied, with a selector index pointer
used to pick which is used by the render function.